Bux fixing

+++ backend/src/controllers/address.controller.ts
@@ -1,6 +1,17 @@
 import { Request, Response, NextFunction } from "express";
-import { deleteAddressService, getAddressByIdService, getAddressService, getCitesServices, getProvincesServices, syncRajaOngkirCitiesService, syncRajaOngkirProvincesService, updateAddressService, userAddressService } from "../services/address.service";
 import { createCustomError } from "../utils/customError";
+import {
+    deleteAddressService,
+    getAddressByIdService,
+    getAddressService,
+    getCitesServices,
+    getProvincesServices,
+    syncRajaOngkirCitiesService,
+    syncRajaOngkirProvincesService,
+    updateAddressService,
+    createUserAddressService
+} from "../services/address.service";
+import { createUpdateAddressSchema } from "../validations/address.validation";
 
 export async function syncRajaOngkirProvincesController(req: Request, res: Response, next: NextFunction) {
     try {
@@ -53,7 +64,8 @@ export async function getCitiesController(req:Request, res:Response, next:NextFu
 
 export async function getAddressController(req:Request, res:Response, next:NextFunction) {
     try {
-        const { userId } = req.body;
+        const userId = req.user?.id;
+        if(!userId) throw createCustomError(401, "Unauthorized")
         const data = await getAddressService(userId);
 
         res.status(200).json({
@@ -64,15 +76,15 @@ export async function getAddressController(req:Request, res:Response, next:NextF
     }
 }
 
-export async function userAddressController(req:Request, res:Response, next:NextFunction) {
+export async function createUserAddressController(req:Request, res:Response, next:NextFunction) {
     try {
-        const { firstName, lastName, provinceId, cityId, address, mainAddress } = req.body;
+        const { firstName, lastName, provinceId, cityId, address, mainAddress } = createUpdateAddressSchema.parse(req.body);
         const userId = req.user?.id
-        if(!userId) throw createCustomError(404, "user not found")
+        if(!userId) throw createCustomError(401, "Unauthorized")
 
-        const data = await userAddressService(firstName, lastName, provinceId, cityId, address, mainAddress, userId);
+        const data = await createUserAddressService(firstName, lastName, provinceId, cityId, address, mainAddress, userId);
 
-        res.status(200).json({
+        res.status(201).json({
             data
         })
     } catch (error) {
@@ -82,22 +94,26 @@ export async function userAddressController(req:Request, res:Response, next:Next
 
 export async function deleteAddressController(req:Request, res:Response, next:NextFunction) {
     try {
-        const { addressId } = req.body;
-        const userId = req.user?.id!
+        const { addressId } = req.params;
+        if (!addressId) throw new Error("Address ID is required");
+
+        const userId = req.user?.id
+        if (!userId) throw new Error("Unauthorized");
         const { message } = await deleteAddressService(addressId, userId);
 
         res.status(200).json({
             message: message
         })
     } catch (error) {
-        
+        console.log(error);
+        next(error);
     }
 }
 
 export async function getAddressByIdController(req:Request, res:Response, next:NextFunction) {
     try {
-        const { addressId } = req.body;
-        const data = await getAddressByIdService(addressId);
+        const { id } = req.params;
+        const data = await getAddressByIdService(id);
 
         res.json({
             data
@@ -109,9 +125,12 @@ export async function getAddressByIdController(req:Request, res:Response, next:N
 
 export async function updateAddressController(req:Request, res:Response, next:NextFunction) {
     try {
-        const { addressId, firstName, lastName, provinceId, cityId, address, mainAddress } = req.body;
+        const { addressId } = req.params;
+        const { firstName, lastName, provinceId, cityId, address, mainAddress } = createUpdateAddressSchema.parse(req.body);
+        const userId = req.user?.id;
+        if(!userId) throw createCustomError(401, "Unauthorized")
 
-        const data = await updateAddressService(addressId, firstName, lastName, provinceId, cityId, address, mainAddress);
+        const data = await updateAddressService(addressId, firstName, lastName, provinceId, cityId, address, mainAddress, userId);
 
         res.status(200).json({
             data
